Rough CMake build improvements.
Adds toplevel CMakeLists to so all subcomponents can be built with a single
cmake invocation. Adds cmake install instructions to the subdirectories to
eliminate need for `cp` statements used in the example `build.sh`. Can now be
built using a call like:

```
git clone https://github.com/ramkikannan/planc
mkdir planc/my_build && cd $_
\# on, say, summit.olcf.ornl.gov: module load gcc openblas armadillo cuda
cmake -DCMAKE_INSTALL_PREFIX=/sw/.testing/belhorn/planc_test \
      -DCMAKE_IGNORE_MKL=1
      -DCMAKE_BUILD_CUDA=1 ../
make
make install
```

This is a pre-requisite to allowing planc to be installable by Spack on OLCF
machines. A work-in-progress spack package is being developed that will install
the pre-requisite openblas and armadillo libraries.

This commit is intended to be a starting point for a refined top-level
CMakeLists.txt and has much room for improvement, particularly in setting
options at the top-level to configure the subdirectory builds. I leave it up to
the primary developers to decide how best to implement a top-level
CMakeLists.txt to meet their needs.